+++ generators/python/core_utilities/sdk/http_client.py
@@ -1,5 +1,6 @@
 import asyncio
 import email.utils
+import json
 import re
 import time
 import typing
@@ -123,6 +124,25 @@ def maybe_filter_request_body(
     return data_content
 
 
+# Abstracted out for testing purposes
+def get_request_body(
+    *,
+    json: typing.Optional[typing.Any],
+    data: typing.Optional[typing.Any],
+    request_options: typing.Optional[RequestOptions],
+    omit: typing.Optional[typing.Any],
+) -> typing.Tuple[typing.Optional[typing.Any], typing.Optional[typing.Any]]:
+    json_body = None
+    data_body = None
+    if data is not None:
+        data_body = maybe_filter_request_body(data, request_options, omit)
+    else:
+        # If both data and json are None, we send json data in the event extra properties are specified
+        json_body = maybe_filter_request_body(json, request_options, omit)
+
+    return json_body, data_body
+
+
 class HttpClient:
     def __init__(
         self,
@@ -166,6 +186,8 @@ class HttpClient:
             else self.base_timeout
         )
 
+        json_body, data_body = get_request_body(json=json, data=data, request_options=request_options, omit=omit)
+
         response = self.httpx_client.request(
             method=method,
             url=urllib.parse.urljoin(f"{base_url}/", path),
@@ -195,8 +217,8 @@ class HttpClient:
                     )
                 )
             ),
-            json=maybe_filter_request_body(json, request_options, omit),
-            data=maybe_filter_request_body(data, request_options, omit),
+            json=json_body,
+            data=data_body,
             content=content,
             files=convert_file_dict_to_httpx_tuples(remove_none_from_dict(files)) if files is not None else None,
             timeout=timeout,
@@ -246,6 +268,8 @@ class HttpClient:
             else self.base_timeout
         )
 
+        json_body, data_body = get_request_body(json=json, data=data, request_options=request_options, omit=omit)
+
         with self.httpx_client.stream(
             method=method,
             url=urllib.parse.urljoin(f"{base_url}/", path),
@@ -275,8 +299,8 @@ class HttpClient:
                     )
                 )
             ),
-            json=maybe_filter_request_body(json, request_options, omit),
-            data=maybe_filter_request_body(data, request_options, omit),
+            json=json_body,
+            data=data_body,
             content=content,
             files=convert_file_dict_to_httpx_tuples(remove_none_from_dict(files)) if files is not None else None,
             timeout=timeout,
@@ -327,6 +351,8 @@ class AsyncHttpClient:
             else self.base_timeout
         )
 
+        json_body, data_body = get_request_body(json=json, data=data, request_options=request_options, omit=omit)
+
         # Add the input to each of these and do None-safety checks
         response = await self.httpx_client.request(
             method=method,
@@ -357,8 +383,8 @@ class AsyncHttpClient:
                     )
                 )
             ),
-            json=maybe_filter_request_body(json, request_options, omit),
-            data=maybe_filter_request_body(data, request_options, omit),
+            json=json_body,
+            data=data_body,
             content=content,
             files=convert_file_dict_to_httpx_tuples(remove_none_from_dict(files)) if files is not None else None,
             timeout=timeout,
@@ -407,6 +433,8 @@ class AsyncHttpClient:
             else self.base_timeout
         )
 
+        json_body, data_body = get_request_body(json=json, data=data, request_options=request_options, omit=omit)
+
         async with self.httpx_client.stream(
             method=method,
             url=urllib.parse.urljoin(f"{base_url}/", path),
@@ -436,8 +464,8 @@ class AsyncHttpClient:
                     )
                 )
             ),
-            json=maybe_filter_request_body(json, request_options, omit),
-            data=maybe_filter_request_body(data, request_options, omit),
+            json=json_body,
+            data=data_body,
             content=content,
             files=convert_file_dict_to_httpx_tuples(remove_none_from_dict(files)) if files is not None else None,
             timeout=timeout,