+++ generators/python/sdk/CHANGELOG.md
@@ -5,9 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.13.1-rc0] - 2024-06-20
+
+- Fix: the Python SDK now does not send additional properties via JSON or data if the request is leveraging the other field.
+
 ## [2.13.0-rc0] - 2024-06-20
 
-- Upgrade: The Python SDK generator is now upgraded to IR V49. 
+- Upgrade: The Python SDK generator is now upgraded to IR V49.
 
 ## [2.12.0-rc0] - 2024-06-25